requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
anyio==3.6.1
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
async-timeout==4.0.2
attrs==22.1.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blessed==1.19.1
blis==0.9.1
brotlipy==0.7.0
catalogue==2.0.8
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==2.1.1
click==8.1.3
confection==0.0.1
contourpy==1.0.5
cycler==0.11.0
cymem==2.0.6
datasets==2.4.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
entrypoints==0.4
executing==1.0.0
fastcore==1.5.29
fastdownload==0.0.7
fastjsonschema==2.16.1
fastprogress==1.0.3
fastrlock==0.8.1
filelock==3.8.0
fonttools==4.37.2
frozenlist==1.3.1
fsspec==2022.8.2
gpustat==1.0.0
graphviz==0.20.1
huggingface-hub==0.9.1
idna==3.4
imageio==2.22.4
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==6.1.1
ipykernel==6.15.3
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==7.7.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.10
jsonschema==4.16.0
jupy==0.2.2
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter_client==7.3.5
jupyterlab==3.4.7
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.1
jupyterlab_server==2.15.1
kiwisolver==1.4.4
langcodes==3.3.0
llvmlite==0.39.1
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.0
matplotlib-inline==0.1.6
mistune==2.0.4
multidict==6.0.2
multiprocess==0.70.13
murmurhash==1.0.8
nbclassic==0.4.3
nbclient==0.6.8
nbconvert==7.0.0
nbformat==5.5.0
nest-asyncio==1.5.5
networkx==2.8.8
notebook==6.4.12
notebook-shim==0.1.0
numba==0.56.4
numpy==1.23.3
nvidia-ml-py==11.495.46
packaging==21.3
pandas==1.4.4
pandocfilters==1.5.0
parso==0.8.3
path==16.6.0
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
preshed==3.0.7
prometheus-client==0.14.1
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==9.0.0
pycparser==2.21
pydantic==1.9.2
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
pytorch-wavelets==1.3.0
pytz==2022.2.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==24.0.0
regex==2022.9.13
requests==2.31.0
responses==0.18.0
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.9.1
Send2Trash==1.8.0
sentencepiece==0.1.97
sigpy==0.1.26
six==1.16.0
smart-open==5.2.1
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
stack-data==0.5.0
terminado==0.15.0
thinc==8.1.1
threadpoolctl==3.1.0
tifffile==2022.10.10
tinycss2==1.1.1
tokenizers==0.12.1
tomli==2.0.1
torch==1.12.1
torchaudio==0.12.1
torchkbnufft==1.4.0
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
transformers==4.22.0
typer==0.4.2
typing_extensions==4.9.0
urllib3==1.26.12
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
widgetsnbextension==3.6.1
xxhash==3.0.0
yarl==1.8.1
zipp==3.8.1
